+++ CHANGES.rst
@@ -7,6 +7,10 @@ Changelog
 - Fix some testes.
   [rafaelbco]
 
+- Fix migrator for AT-based types that got broken in 1.5.8 release and add
+  an option to remove the content of the non-blob field during migration to
+  not end up having stale data in the ZODB
+  [fRiSi]
 
 1.5.10 (2014-04-16)
 -------------------